requirements.txt
numpy
pandas
faiss-cpu
sentence-transformers
accelerate>=0.26.0
groq
hugchat
langdetect 
deep-translator
streamlit>=1.20.0
torch
gtts
SpeechRecognition
streamlit==1.25.0
streamlit-webrtc==0.43.0
pydub

